# openapi/profile.yaml
openapi: 3.0.0
info:
  title: Profile API
  version: 1.0.0
paths:
  /api/profiles:
    post:
      summary: Create a new profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfile"
      responses:
        "201":
          description: Profile created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
  /api/profiles/{id}:
    get:
      summary: Get a profile by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: mode
          in: query
          required: false
          schema:
            type: string
            enum: [edit]
          description: When set to 'edit', returns profile data in CreateProfile format including full configuration from R2
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Profile"
                  - $ref: "#/components/schemas/CreateProfile"
    put:
      summary: Update a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfile"
      responses:
        "204":
          description: Profile updated successfully
    delete:
      summary: Delete a profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Profile deleted successfully
  /api/profiles/{id}/singbox:
    get:
      summary: Get the SingBox profile URL
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    format: uri
components:
  schemas:
    Profile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        tags:
          type: array
          items:
            type: string
        createdBy:
          type: string
          format: uuid
        outbounds:
          type: array
          items:
            type: integer
        rule_sets:
          type: array
          items:
            type: integer
    CreateProfile:
      type: object
      properties:
        name:
          type: string
        tags:
          type: array
          items:
            type: string
        outbounds:
          type: array
          items:
            type: integer
        route:
          $ref: "#/components/schemas/Route"
        dns:
          $ref: "#/components/schemas/Dns"
        inbounds:
          type: array
          items:
            $ref: "#/components/schemas/Inbound"
    UpdateProfile:
      $ref: "#/components/schemas/CreateProfile"
    Route:
      type: object
      properties:
        rule_set:
          type: array
          items:
            type: integer
    Dns:
      type: object
    Inbound:
      type: object